% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcloss.R
\name{CTC_loss}
\alias{CTC_loss}
\title{CTC_loss() Function}
\usage{
CTC_loss(output_prob, label, blank_idx)
}
\arguments{
\item{output_prob:}{output probability matrix of model, b: ground truth label, blank_idx: index of blank token}
}
\value{
CTC loss
}
\description{
This function calculate CTC loss.
}
\keyword{ctcloss}
